store workout incline as a float instead of integer

+++ db/migrate/20140716005348_create_workouts.rb
@@ -5,7 +5,7 @@ class CreateWorkouts < ActiveRecord::Migration
       t.belongs_to :activity
       t.integer :time #seconds
       t.integer :distance #feet
-      t.integer :incline #percentage
+      t.float :incline #percentage
       t.date :date
       t.string :note, limit: 2048
       t.timestamps
